Refined ContentDirectory Implementation. Updated interfaces to support offset and limit (with backwards compatibility).
Implemented changes so client will work also, no paging yet.

Have to disable NativeTv plugin to get working as it is causing issues with the database.

Note: All UPnP custom functions for MediaPortal 2 are now prefixed with X_MediaPortal_ ( really should be X_team.mediaportal.com_ ) so it's no long compatible with eariler versions of MediaPortal 2. This is done to conform with the UPnP spec.